+++ ios/Runner/AppDelegate.m
@@ -6,6 +6,9 @@
 
 - (BOOL)application:(UIApplication *)application
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
+  if (@available(iOS 10.0, *)) {
+    [UNUserNotificationCenter currentNotificationCenter].delegate = (id<UNUserNotificationCenterDelegate>) self;
+  }
   [GMSServices provideAPIKey:@"AIzaSyD7fSvAPIz-31x_pMPJ3hR3xMcP8WjJxJ0"];
   [GeneratedPluginRegistrant registerWithRegistry:self];
   // Override point for customization after application launch.