+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Jul 26, 2020
+
+- Added support for Firebase messaging, including access to the token from the drawer menu 
+  (click on the user details in the header)
+- Bump version to 1.3.1+5
+
 ## Jul 25, 2020
 
 - Added a drawer menu, moved logout into menu, added menu options to refresh tokens and get user info, as well as to view